+++ i-tapp-client-main/src/components/layouts/protected/student/index.tsx
@@ -11,7 +11,7 @@ const StudentLayout = ({ children }: { children: ReactNode }) => {
   return (
     <>
       <Header link={studentNavLinks} />
-      <main className="h-screen bg-[#F0F0F5]">{children}</main>
+      <main className="h-screen bg-muted dark:bg-background">{children}</main>
       <OnboardingTour role="student" />
       <AppOnly>
         <AppTabBar role="student" />